+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### v1.1.6  (2014-02-02)
+* * *    
+* Support for refund invoice and transaction.
+
+* Support for creating plans & addons on the fly via API.
+
+* Support for handling double data type params handled. Eg., download_penality param for create new plan.
+
 ### v1.1.5  (2014-01-18)  
 * * *    
 * Adding object that represent comments resource. Now comments can be added to the entities - Subscription, Invoice, Transaction, Plan, Addon & Coupon.